requirements.txt
# opencv-python==4.5.5.62
opencv-python
tqdm
pytorch-msssim
timm
tensorboard
tensorboardx